Move functions to get expected Xray results into XraySteps since they are not specific to Splunk
